% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{biodbExpasy-package}
\alias{biodbExpasy}
\alias{biodbExpasy-package}
\title{biodbExpasy: biodbExpasy, a library for connecting to Expasy ENZYME database.}
\description{
The biodbExpasy library provides access to Expasy ENZYME database, using biodb package framework. It allows to retrieve entries by their accession number. Web services can be accessed for searching the database by name or comments.
}
\details{
See vignette biodbExpasy:\preformatted{vignette('biodbExpasy', package='biodbExpasy')
}
}
\seealso{
\link{ExpasyEnzymeConn}.
}
\author{
\strong{Maintainer}: Pierrick Roger \email{pierrick.roger@cea.fr} (\href{https://orcid.org/0000-0001-8177-4873}{ORCID})

}
