% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{biodbExpasy-package}
\alias{biodbExpasy}
\alias{biodbExpasy-package}
\title{biodbExpasy: biodbExpasy, a library for connecting to Expasy ENZYME database.}
\description{
The biodbExpasy library is an extension of the biodb framework package, that provides access to Expasy ENZYME database. It allows to retrieve entries by their accession number, and run specific web services.
}
\details{
See vignette biodbExpasy:\preformatted{vignette('biodbExpasy', package='biodbExpasy')
}
}
\seealso{
\link{ExpasyEnzymeConn}.
}
\author{
\strong{Maintainer}: Pierrick Roger \email{pierrick.roger@cea.fr}

}
